Move SmartBulb into SmartDevice
